[stack-manager] Fix illegal read actions
