+++ src/com/haskforce/tooling/hpack/PackageYamlQuery.scala
@@ -1,5 +1,6 @@
 package com.haskforce.tooling.hpack
 
+import com.intellij.openapi.application.ReadAction
 import com.intellij.psi.PsiFile
 import org.jetbrains.yaml.YAMLUtil
 import org.jetbrains.yaml.psi.YAMLFile
@@ -8,22 +9,26 @@ import org.jetbrains.yaml.psi.YAMLFile
 object PackageYamlQuery {
 
    def getName(x: YAMLFile): Either[Throwable, String] = {
-     Option(YAMLUtil.getQualifiedKeyInFile(x, "name"))
-       .map(_.getValueText)
-       .toRight(new NoSuchElementException(s"'name' field in ${x.getName}"))
+     ReadAction.compute(() => {
+       Option(YAMLUtil.getQualifiedKeyInFile(x, "name"))
+         .map(_.getValueText)
+         .toRight(new NoSuchElementException(s"'name' field in ${x.getName}"))
+     })
    }
 
   def getTopLevelDeps(packageYaml: PsiFile): Option[List[String]] = {
-    val res =
-      packageYaml.getText.split('\n')
-        .iterator
-        .dropWhile(_.trim != "dependencies:")
-        .drop(1)
-        .map(_.trim)
-        .takeWhile(s => s == "" || s.startsWith("-"))
-        .map(_.stripPrefix("-").trim)
-        .filter(_.nonEmpty)
-        .toList
-    if (res.isEmpty) None else Some(res)
+    ReadAction.compute(() => {
+      val res =
+        packageYaml.getText.split('\n')
+          .iterator
+          .dropWhile(_.trim != "dependencies:")
+          .drop(1)
+          .map(_.trim)
+          .takeWhile(s => s == "" || s.startsWith("-"))
+          .map(_.stripPrefix("-").trim)
+          .filter(_.nonEmpty)
+          .toList
+      if (res.isEmpty) None else Some(res)
+    })
   }
 }
